Fix bug with double-digit question indexes
There was a bug with the software where when the number of questions
reached 10, question 10 would actually result in the display of
question 1. This was because the program finds the program index from
the current element's ID, where the conversion from string to integer
was done using parseInt() with a parameter referencing a specific
character of the string, being the number. However, since the reference
was one character, the program would ignore double-digit (or more) IDs.

Therefore, the id "qu10" would be abbreviated as index 1 as the program
takes in the 3rd character and processes it, ignoring the 4th character.